requirements.txt
requests
pandas
pytest
pytest-mock
pylint
coverage
python-dotenv
boto3
pymssql
altair
streamlit